fix(p0-2): split dispatch_hook.rs into module to satisfy file_size_invariant
CI on r1 caught `tests/file_size_invariant.rs` failing with
`src/mcp/handlers/dispatch_hook.rs: 716 LOC (max 700)`.

Per lead's recommendation (Option A in r1 dispatch), convert the file into
a directory module:

- `src/mcp/handlers/dispatch_hook/mod.rs` — production code (120 LOC)
- `src/mcp/handlers/dispatch_hook/tests.rs` — test module body (591 LOC)

The invariant in `tests/file_size_invariant.rs` only iterates the top level
of `src/mcp/handlers/` and skips files named exactly `tests.rs`, so the
new `dispatch_hook/` subdirectory is excluded entirely from the LOC check.
The `tests.rs` filename also matches the existing skip rule for the shared
`src/mcp/handlers/tests.rs` (2562 LOC) precedent.

No production-code change. Test contents unchanged (just dedented after
moving out of the inline `mod tests { ... }` block — rustfmt handled it).

Verification:
- cargo test --bin agend-terminal dispatch_hook → 13/13 pass
- cargo test --test file_size_invariant → 1/1 pass
- cargo fmt + cargo clippy --all-targets -- -D warnings → clean
